Basic Flying Ships
The ship can now be driven by the owner after it has been fully constructed.
Currently there is no way to dismount or park the ship for the player, although the functionality is there on the server.
Clients are given network control, so velocity checks from the server must be implemented.