Selection Boxes
Added the "Select" method to the GUI module in ReplicatedStorage to give all buttons a clean look when hovered over, without using AutoButtonColor.
Outlines any gui object in white as long as it has a visible background and matches its shape.
Slides if changed quickly or fades in/out.
Adjusts size and position dynamically.